🚚 Rename: Rename api and move interfaces
